@@ -11,8 +11,8 @@
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "buffer": "^6.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-html-parser": "^2.0.2",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.0",
@@ -56,5 +56,5 @@
     "postcss": "^8.2.13",
     "css-select": "^2.1.0",
     "nth-check": "^2.0.1"
-}
+  }
 }
